Move fitler to awx/common. Update imports for Jobs list.
