+++ frontend/awx/common/useDynamicFilters.tsx
@@ -1,13 +1,13 @@
 import { useCallback, useMemo } from 'react';
 import { useTranslation } from 'react-i18next';
-import { IToolbarFilter, ToolbarFilterType } from '../../../../framework';
-import { useOptions } from '../../../common/crud/useOptions';
-import { AwxItemsResponse } from '../../common/AwxItemsResponse';
-import { awxAPI } from '../../common/api/awx-utils';
-import { QueryParams } from '../../common/useAwxView';
-import { requestGet } from '../../../common/crud/Data';
-import { DateRangeFilterPresets } from '../../../../framework/PageToolbar/PageToolbarFilters/ToolbarDateRangeFilter';
-import { ActionsResponse, OptionsResponse } from '../../interfaces/OptionsResponse';
+import { IToolbarFilter, ToolbarFilterType } from '../../../framework';
+import { DateRangeFilterPresets } from '../../../framework/PageToolbar/PageToolbarFilters/ToolbarDateRangeFilter';
+import { requestGet } from '../../common/crud/Data';
+import { useOptions } from '../../common/crud/useOptions';
+import { AwxItemsResponse } from '../common/AwxItemsResponse';
+import { awxAPI } from '../common/api/awx-utils';
+import { QueryParams } from '../common/useAwxView';
+import { ActionsResponse, OptionsResponse } from '../interfaces/OptionsResponse';
 
 interface DynamicToolbarFiltersProps {
   /** API endpoint for the options that generated the filters */
